+++ FluentFTP.Tests/Unit/CustomTests.cs
@@ -12,10 +12,10 @@ namespace FluentFTP.Tests.Unit {
 	public class CustomTests {
 
 		private FtpClient NewTestClient() {
-			return new FtpClient("", 21, "", "");
+			return new FtpClient("1.2.3.4", new NetworkCredential("user", "pass"));
 		}
 		private async Task<AsyncFtpClient> NewTestAsyncClient() {
-			return new AsyncFtpClient("", 21, "", "");
+			return new AsyncFtpClient("1.2.3.4", new NetworkCredential("user", "pass"));
 		}
 
 		[Fact]